chore(ssa): Refactor `array_set_rc_invariant` into `rc_invariant::array_set` (#13162)
