+++ compiler/noirc_evaluator/src/ssa/validation/mod.rs
@@ -28,9 +28,9 @@ use acvm::{
 use itertools::Itertools;
 use rustc_hash::{FxHashMap as HashMap, FxHashSet as HashSet};
 
-#[cfg(debug_assertions)]
-pub(crate) mod array_set_rc_invariant;
 pub(crate) mod dynamic_array_indices;
+#[cfg(debug_assertions)]
+pub(crate) mod rc_invariant;
 
 use crate::ssa::{
     ir::{
